requirements.txt
turbopuffer
voyageai
openai
requests
pydantic
python-dotenv
pyyaml
sentence-transformers
